remove isDescending from Query interface, since it is only actually settable and usable by TimeseriesQuery
