+++ processing/src/main/java/org/apache/druid/query/select/SelectQuery.java
@@ -111,12 +111,6 @@ public class SelectQuery implements Query<Object>
   }
 
   @Override
-  public boolean isDescending()
-  {
-    throw new RuntimeException(REMOVED_ERROR_MESSAGE);
-  }
-
-  @Override
   public Ordering<Object> getResultOrdering()
   {
     throw new RuntimeException(REMOVED_ERROR_MESSAGE);
